feat: cross_entropy_loss impl

+++ include/loss.h
@@ -3,7 +3,7 @@
 
 #include "tensor.h"
 
-int cross_entropy_loss(Tensor* predictions, Tensor* targets);
-int cross_entropy_loss_backward(Tensor* predictions, Tensor* targets, Tensor* grad);
+double cross_entropy_loss(Tensor* predictions, Tensor* targets);
+double cross_entropy_loss_backward(Tensor* predictions, Tensor* targets, Tensor* grad);
 
 #endif
